# doc/api.yaml
openapi: 3.0.3
info:
  title: Road To Green Island Api
  description: Api to help Malagasy people to get to know our green spaces and our various species
  version: latest
servers:
  - url: 'http://localhost:8080'
paths:
...
  /green-spaces:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getGreenSpaces
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/page_size"
      security:
        - bearerAuth: [ ]
      tags:
        - Green_Spaces
      responses:
        '200':
          description: list of green spaces
          content:
            'application/json':
              schema:
                type: object
                allOf:
                  - $ref: "#/components/schemas/PagedResponse"
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/GreenSpace"
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /green-spaces/{gsp_id}:
    parameters:
      - name: gsp_id
        required: true
        schema:
          type: string
        in: path
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Green_Spaces
      operationId: getGreenSpace
      responses:
        '200':
          description: green space identified by its id
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/GreenSpace"
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      security:
        - bearerAuth: [ ]
      operationId: crupdateGreenSpace
      tags:
        - Green_Spaces
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/GreenSpace"
      responses:
        '200':
          description: crupdated greenspace
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/GreenSpace"

        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  parameters:
    page:
      name: page
      in: query
      description: page number
      schema:
        type: integer
        minimum: 1
        default: 1
    page_size:
      name: page_size
      in: query
      description: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 30
  schemas:
...
    GreenSpaceProfile:
      type: object
      properties:
        area:
          type: number
          format: double
        description:
          type: string
          description: description of the green space
        location:
          $ref: "#/components/schemas/Location"
        ticket_price:
          type: number
          format: double
    Location:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        name:
          type: string
        address:
          type: string
    GreenSpace:
      type: object
      properties:
        id:
          type: string
        user_id:
          description: owner or representative id
          type: string
        profile:
          $ref: "#/components/schemas/GreenSpaceProfile"
      required:
        - id
...
    PagedResponse:
      type: object
      properties:
        page_size:
          type: integer
        page_number:
          type: integer
        has_previous:
          type: boolean
        count:
          type: integer
    #############################################
    ###############  Exceptions #################
    #############################################
    ExceptionModel:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: InternalServerException
        message: Unexpected error